fix: inconsistent on register node after layout change
